test(usability): Harden stats schemas against float inputs
History-Category: Changed
History-Impact: 0.1
History-Confidence: 1
History-Validation: passed
History-Journal: true

+++ src/adapters/mysql/tools/stats/comparative/schemas.ts
@@ -139,7 +139,7 @@ export const HistogramSchema = z.preprocess(
   z.object({
     table: z.string().min(1, "table is required"),
     column: z.string().min(1, "column is required"),
-    buckets: z.coerce.number().min(1).default(16),
+    buckets: z.coerce.number().int().min(1).default(16),
     update: z.boolean().default(false),
   })
 );
